+++ src/main/java/com/communitystreet/service/ProfileService.java
@@ -1,13 +1,12 @@
 package com.communitystreet.service;
 
 import com.communitystreet.domain.Activity;
-import com.communitystreet.domain.Community;
 import com.communitystreet.domain.User;
 
 import java.util.List;
 
 public interface ProfileService {
-    User getUser(int stuNumber);
+    User getUser(long stuNumber);
     int updateUser(User user);
 //    List<Community> getCommunities(int stuNumber);
 //    boolean createCommunity(Community community);
